io_queue costs shouldn't be zero for large bandwidth/iops
This patch fixes a case in which given large enough
iops/bandwitdh configured in io-properties, the requests costs
would show up 0.

The bug is in internal::request_tokens, for bigger configured iops and
bandwidth, disk_req_write_to_read_multiplier and
disk_blocks_write_to_read_multiplier end up < 1, so casting them to
unsigned would make them 0, so in the end costs end up 0.
They should end up 0 anyway the closer you get to uint64_t max, but not
that soon.

Fixes #2816

+++ src/core/io_queue.cc
@@ -880,8 +880,8 @@ stream_id io_queue::request_stream(io_direction_and_length dnl) const noexcept {
 
 double internal::request_tokens(io_direction_and_length dnl, const io_queue::config& cfg) noexcept {
     struct {
-        unsigned weight;
-        unsigned size;
+        double weight;
+        double size;
     } mult[2];
 
     mult[io_direction_write] = {
